feat: implement Notification and Promotion controllers, add Jackson and OpenAPI configurations, and enhance messaging services.

+++ src/main/java/com/furniro/MessageService/database/entity/Subscription.java
@@ -12,11 +12,11 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class  Subscription {
+public class Subscription {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotBlank(message = "Họ và tên không được để trống")
     @Size(min = 2, max = 100, message = "Họ và tên phải từ 2-100 ký tự")
