+++ src/main/java/com/furniro/MessageService/database/repository/PromotionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.furniro.MessageService.database.entity.Promotion;
 
-public interface PromotionRepository extends JpaRepository<Promotion, Long> {
+public interface PromotionRepository extends JpaRepository<Promotion, Integer> {
 
 }
