+++ src/main/java/com/furniro/MessageService/database/repository/SubscriptionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.furniro.MessageService.database.entity.Subscription;
 
-public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
+public interface SubscriptionRepository extends JpaRepository<Subscription, Integer> {
     boolean existsByEmail(String email);
 
     Page<Subscription> findAll(Pageable pageable);
